% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedbib.R
\name{add_document}
\alias{add_document}
\title{Add reference document to project}
\usage{
add_document(bibtex_key,
  overwrite_references = get_config_value("overwrite_references"),
  link_references = get_config_value("link_references"),
  local_references_dir = get_config_value("local_references_dir"), ...)
}
\arguments{
\item{bibtex_key}{The bibtex key for the citation and the file name.}

\item{overwrite_references}{TRUE/FALSE. If TRUE, overwrite existing reference documents.}

\item{link_references}{TRUE/FALSE. If TRUE, make a link to the document in the library. If FALSE, copy the document.}

\item{local_references_dir}{A relative path to where to store the file. If the directory does not already exist it will be created.}

\item{...}{Arguments passed on to \code{library_path}
\describe{
  \item{}{}
}}
}
\description{
Add a local copy or link to the document referenced by a bibtex key in a folder.
The folder the copy of the document is stored in is defined by the option "local_references_dir" in the user's configuration file or in \link{ref_options}.
}
