% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedbib.R
\name{document_path}
\alias{document_path}
\title{Get path to document associated with a citation}
\usage{
document_path(bibtex_key, ...)
}
\arguments{
\item{bibtex_key}{The bibtex key for the citation and the file name.}

\item{...}{Arguments passed on to \code{library_path}
\describe{
  \item{}{}
}}
}
\description{
Get the full path to a document (e.g. pdf) associated with a bibtex entry.
The document should be named the same as the bibtex key, besides for the file extension
}
