% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_file_in_parent}
\alias{find_file_in_parent}
\title{Find file in a parent directory}
\usage{
find_file_in_parent(file_name, dir_path = getwd(), regex = FALSE)
}
\arguments{
\item{file_name}{The name of the file.}

\item{dir_path}{The directory to start the search from. Defaults to current
working directory.}

\item{regex}{TRUE/FALSE. Is the \code{file_name} option a regular expression?}
}
\description{
Search all parent directories of a given directory and return the path to the
first file found with a given name.
}
\keyword{internal}
