% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_config}
\alias{get_config}
\title{Get configuration settings}
\usage{
get_config(dir_path = getwd(),
  config_name = ref_options$get("config_file_name"))
}
\arguments{
\item{dir_path}{The directory to start the search from. Defaults to current
working directory.}

\item{config_name}{The name of the configuration file. The default name of the configuration file is defined by the
global constant variable \code{config_file_name} in \link{ref_options}.}
}
\description{
Get settings stored in a configuration file
}
\keyword{internal}
