% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{get_config_value}
\alias{get_config_value}
\title{Get a configuration setting}
\usage{
get_config_value(value, allow_global_default = TRUE, ...)
}
\arguments{
\item{value}{The name of the setting to get}

\item{allow_global_default}{TRUE/FALSE. If TRUE and a value cannot be found
in a configuration file and the same value is a global package option
defined in \link{ref_options}, then use the global pacakge option.}

\item{...}{Arguments passed on to \code{get_config}
\describe{
  \item{dir_path}{The directory to start the search from. Defaults to current
working directory.}
  \item{config_name}{The name of the configuration file. The default name of the configuration file is defined by the
global constant variable \code{config_file_name} in \link{ref_options}.}
}}
}
\description{
Get a setting stored in a configuration file
}
\keyword{internal}
