% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmanager.R
\name{library_path}
\alias{library_path}
\title{Find path to library}
\usage{
library_path(...)
}
\arguments{
\item{...}{Arguments passed on to \code{get_config}
\describe{
  \item{dir_path}{The directory to start the search from. Defaults to current
working directory.}
  \item{config_name}{The name of the configuration file. The default name of the configuration file is defined by the
global constant variable \code{config_file_name} in \link{ref_options}.}
}}
}
\description{
Searchs for the path to the library folder containing documents assocaited
with bibtex citations in the bibliography by looking for a configuration file
in a parent directory.
}
