% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedbib.R
\name{link_all_references}
\alias{link_all_references}
\title{Link all documents referenced in R markdown}
\usage{

  link_all_references(local_references = get_config_value("local_references"))
}
\arguments{
\item{local_references}{TRUE/FALSE. If TRUE, then a copy/link of the document
is added in a folder in the current working directory. See \link{add_document}
for details. If FALSE, the document in the users' library is linked to.}
}
\description{
Add links to local copies or link to the documents referenced by a bibtex
key. The folder the copy of the document is stored in is defined by the
option "local_references_dir" in the user's configuration file or in
\link{ref_options}.
}
