% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmanager.R
\name{link_document}
\alias{link_document}
\title{Link a reference document in R markdown}
\usage{
link_document(bibtex_key, link_text = NULL,
  local_references = get_config_value("local_references"), ...)
}
\arguments{
\item{bibtex_key}{The bibtex key for the citation and the file name.}

\item{link_text}{The text that will be shown and can be clicked on. By
default, text is made from information in the bibtex file.}

\item{local_references}{TRUE/FALSE. If TRUE, then a copy/link of the document
is added in a folder in the current working directory. See \link{add_document}
for details. If FALSE, the document in the users' library is linked to.}

\item{...}{Arguments passed on to \code{add_document}
\describe{
  \item{bibtex_key}{The bibtex key for the citation and the file name.}
  \item{overwrite_references}{TRUE/FALSE. If TRUE, overwrite existing reference documents.}
  \item{link_references}{TRUE/FALSE. If TRUE, make a link to the document in the library. If FALSE, copy the document.}
  \item{local_references_dir}{A relative path to where to store the file. If the directory does not already exist it will be created.}
}}
}
\description{
Add a link to a local copy or link to the document referenced by a bibtex
key. The folder the copy of the document is stored in is defined by the
option "local_references_dir" in the user's configuration file or in
\link{ref_options}.
}
