% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_options.R
\name{new_default_options}
\alias{new_default_options}
\title{Returns a list of functions assocaited with a stored set of options.}
\usage{
new_default_options(value = list())
}
\arguments{
\item{value}{The list of possible options.}
}
\description{
Returns a list of functions assocaited with a stored set of options.
}
\section{Attribution}{


This code is copied from the code handling options in \link{knitr}.
}

\keyword{internal}
