% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parent_dir_paths}
\alias{parent_dir_paths}
\title{Get paths to all parent directories}
\usage{
parent_dir_paths(dir_path = getwd())
}
\arguments{
\item{dir_path}{The directory to start the search from. Defaults to current
working directory.}
}
\description{
Returns a vector with the full paths to all parent directories of a directory.
}
\keyword{internal}
