% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{read_config_file}
\alias{read_config_file}
\title{Read a configuration file}
\usage{
read_config_file(file_path)
}
\arguments{
\item{file_path}{The full path to the file to read.}
}
\description{
Read a yaml configuartion file
}
\keyword{internal}
