% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_options.R
\docType{data}
\name{ref_options}
\alias{ref_options}
\title{Default options}
\format{An object of class \code{list} of length 5.}
\usage{
ref_options
}
\description{
The default values for package-wide options.
}
\section{Attribution}{


This code is copied from the code handling options in \link{knitr}.
}

\examples{

# List all options
ref_options$get()

# Get a spcific option
ref_options$get('config_file_name')

# Set one or more options
ref_options$set(config_file_name =  ".my_comstom_config.yml")

}
\keyword{datasets}
